+++ ACP3/Modules/ACP3/Users/Validation/AccountSettingsFormValidation.php
@@ -34,6 +34,7 @@ class AccountSettingsFormValidation extends AbstractUserFormValidation
         $this->validator->addConstraint(Core\Validation\ValidationRules\FormTokenValidationRule::class);
 
         $this->validateUserSettings($formData, $this->settings['language_override']);
+        $this->validateNewPassword($formData, 'new_pwd', 'new_pwd_repeat');
 
         $this->validator->validate();
     }